Rename Home to index. Rename Signup component to SignupForm. Move function to SignupForm.
